backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.19
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
apscheduler==3.10.4
httpx==0.27.0
aiofiles==23.2.1
python-nmap==0.7.1
manuf==1.1.5
bcrypt==4.1.3
